completely replace rendering logic of ShipScheme widget with canvas
