+++ lib/presentation/ship_scheme/widgets/ship_cheme_body.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:hmi_core/hmi_core_app_settings.dart';
 import 'package:sss_computing_client/presentation/core/models/ship_scheme/chart_axis.dart';
 import 'package:sss_computing_client/presentation/core/models/ship_scheme/figure.dart';
-import 'package:sss_computing_client/presentation/ship_scheme_refactored/widgets/figures_test.dart';
-import 'package:sss_computing_client/presentation/ship_scheme_refactored/widgets/ship_scheme.dart';
-import 'package:sss_computing_client/presentation/ship_scheme_refactored/widgets/ship_scheme_options.dart';
+import 'package:sss_computing_client/presentation/ship_scheme/widgets/figures_test.dart';
+import 'package:sss_computing_client/presentation/ship_scheme/widgets/ship_scheme.dart';
+import 'package:sss_computing_client/presentation/ship_scheme/widgets/ship_scheme_options.dart';
 
 ///
 const Set<ShipSchemeOption> _initialOptions = {
