+++ lib/presentation/ship_scheme/widgets/ship_scheme_figure.dart
