+++ lib/presentation/ship_scheme/widgets/ship_scheme_figures.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:sss_computing_client/presentation/core/models/ship_scheme/figure.dart';
-import 'package:sss_computing_client/presentation/ship_scheme_refactored/widgets/ship_scheme_figure.dart';
+import 'package:sss_computing_client/presentation/ship_scheme/widgets/ship_scheme_figure.dart';
 
 ///
 class ShipSchemeFigures extends StatelessWidget {
